Reimplemented header file generation for nbody
